Extract swaption FpML mapping
